+++ src/main/java/mx/edu/utez/fastfood/dao/DishDao.java
@@ -1,10 +1,10 @@
-package mx.edu.utez.fastfood.fastfood.dao;
+package mx.edu.utez.fastfood.dao;
 
-import mx.edu.utez.fastfood.fastfood.exception.DaoException;
-import mx.edu.utez.fastfood.fastfood.model.Category;
-import mx.edu.utez.fastfood.fastfood.model.Dish;
-import mx.edu.utez.fastfood.fastfood.model.Ingredient;
-import mx.edu.utez.fastfood.fastfood.service.MySqlConnection;
+import mx.edu.utez.fastfood.exception.DaoException;
+import mx.edu.utez.fastfood.model.Category;
+import mx.edu.utez.fastfood.model.Dish;
+import mx.edu.utez.fastfood.model.Ingredient;
+import mx.edu.utez.fastfood.service.MySqlConnection;
 
 import java.sql.*;
 import java.util.*;
