+++ src/main/java/mx/edu/utez/fastfood/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package mx.edu.utez.fastfood.fastfood.exception;
+package mx.edu.utez.fastfood.exception;
 
 public class ValidationException extends Exception {
     public ValidationException(String message) {
