+++ src/main/java/mx/edu/utez/fastfood/service/MySqlConnection.java
@@ -1,4 +1,4 @@
-package mx.edu.utez.fastfood.fastfood.service;
+package mx.edu.utez.fastfood.service;
 
 import java.sql.*;
 import java.util.ResourceBundle;
